[server] Migrate all ports-related API calls to WorkspaceService
